move send_notification back to updater.py
